Get offset by name

+++ TypeSystem/TypePartDefinitions/SystemTypeFieldDefinition/SystemTypeFieldDefinition.h
@@ -20,18 +20,18 @@ namespace types {
 
         std::string name_;
         const SystemType* type_;
-        byte offset_;
+        off_t offset_;
         TypeFlags flags_;
 
     public:
-        SystemTypeFieldDefinition(std::string name, const SystemType& type, byte offset, TypeFlags flags)
+        SystemTypeFieldDefinition(std::string name, const SystemType& type, off_t offset, TypeFlags flags)
             : name_(name), type_(&type), offset_(offset), flags_(flags)
         {
 
         }
 
         std::string name() const { return name_; }
-        byte offset() { return offset_; }
+        off_t offset() { return offset_; }
         const SystemType& type() const { return *type_;}
         TypeFlags flags() { return flags_; }
